requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Environment & Configuration
python-dotenv==1.0.0

# OpenAI API
openai==1.6.1

# Email Processing
imapclient==3.0.1
email-validator==2.1.0
BeautifulSoup4==4.12.2

# Telegram Bot
python-telegram-bot==20.7

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3